# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.ObserverTelemetry:
    properties:
      interval:
        type: string
      points:
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ObserverTelemetryPoint'
        type: array
      range:
        type: string
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.ObserverTelemetryPoint:
    properties:
      airtimeRxPct:
        type: number
      airtimeTxPct:
        type: number
      batteryMv:
        type: integer
      noiseFloorDb:
        type: number
      queueLength:
        type: integer
      receiveErrors:
        type: integer
      t:
        description: epoch ms
        type: integer
      uptimeSeconds:
        type: integer
    type: object
...
  ? github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_PublicAdvertObservation
  : properties:
      hasMore:
        type: boolean
      items:
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PublicAdvertObservation'
        type: array
      nextCursor:
        type: integer
    type: object
...
  ? github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_PublicObserverSummary
  : properties:
      hasMore:
        type: boolean
      items:
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PublicObserverSummary'
        type: array
      nextCursor:
        type: integer
    type: object
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.PublicAdvertObservation:
    properties:
      heardAt:
        type: integer
      hopCount:
        type: integer
      iata:
        type: string
      id:
        type: integer
      nodeName:
        type: string
      nodePublicKey:
        type: string
      packetHash:
        type: string
      payloadType:
        type: integer
      payloadTypeName:
        type: string
      rssi:
        type: integer
      snr:
        type: number
    type: object
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.PublicObserver:
    properties:
      batteryLevel:
        type: number
      brokers:
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PublicObserverBroker'
        type: array
      displayName:
        type: string
      firmwareBuild:
        type: string
      firmwareVersion:
        type: string
      firstSeen:
        type: integer
      hardwareModel:
        type: string
      iata:
        type: string
      id:
        type: string
      lastSeen:
        type: integer
      lastStatusAt:
        type: integer
      observationCount:
        type: integer
      observerType:
        type: string
      publicKey:
        type: string
      radio:
        type: string
      radioBwKhz:
        type: number
      radioCr:
        type: integer
      radioFreqMhz:
        type: number
      radioSf:
        type: integer
      scopes:
        items:
          type: string
        type: array
      softwareVersion:
        type: string
      status:
        type: string
      uptimeSeconds:
        type: integer
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.PublicObserverBroker:
    properties:
      lastPacketAt:
        description: epoch ms, 0 if none
        type: integer
      lastSeenAt:
        description: epoch ms
        type: integer
      name:
        type: string
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.PublicObserverSummary:
    properties:
      displayName:
        type: string
      iata:
        type: string
      id:
        type: string
      observerType:
        type: string
      radio:
        type: string
      scopes:
        items:
          type: string
        type: array
      status:
        type: string
    type: object
...
  internal_api_handlers.APIError:
    properties:
      code:
        description: e.g. "not_found", "bad_request"
        type: string
      message:
        description: e.g. "channel not found"
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: MeshCore Beacon
    url: https://github.com/MeshCore-Beacon/beacon-server
  description: MeshCore network observation backend. Ingests LoRa packets from MQTT
    brokers, stores in PostgreSQL, and streams live events via WebSocket.
  license:
    name: AGPL-3-or-later
  termsOfService: https://github.com/MeshCore-Beacon/beacon-server
  title: MeshCore Beacon API
  version: 1.6.0
paths:
...
  /observers:
    get:
      parameters:
      - description: Filter by single IATA code (case-insensitive)
        in: query
        name: iata
        type: string
      - description: Filter by multiple IATA codes, comma-separated e.g. YVR,YYJ
        in: query
        name: iatas
        type: string
      - description: Filter by region ID, expands to member IATAs
        in: query
        name: regionId
        type: integer
      - description: Filter by region slug, expands to member IATAs
        in: query
        name: region
        type: string
      - description: Filter by observer type (e.g. meshcoretomqtt, meshcore-ha)
        in: query
        name: type
        type: string
      - description: Filter by broker name
        in: query
        name: broker
        type: string
      - description: Filter by status (online or offline)
        in: query
        name: status
        type: string
      - description: Partial case-insensitive display name match
        in: query
        name: name
        type: string
      - description: 'Filter by transport scope name e.g. %23bc (URL-encoded #bc)'
        in: query
        name: scope
        type: string
      - description: last_seen epoch ms of last item for pagination
        in: query
        name: cursor
        type: integer
      - description: Max results (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_PublicObserverSummary'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: List observers
      tags:
      - Observers
  /observers/{observerId}:
    get:
      parameters:
      - description: Observer UUID
        in: path
        name: observerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PublicObserver'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Get observer detail
      tags:
      - Observers
  /observers/{observerId}/adverts:
    get:
      parameters:
      - description: Observer UUID
        in: path
        name: observerId
        required: true
        type: string
      - description: Observation ID of last item for pagination
        in: query
        name: cursor
        type: integer
      - description: Max results (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_PublicAdvertObservation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: List advert packets heard by an observer
      tags:
      - Observers
  /observers/{observerId}/telemetry:
    get:
      parameters:
      - description: Observer UUID
        in: path
        name: observerId
        required: true
        type: string
      - description: Duration window e.g. 24h, 48h, 168h (default 24h)
        in: query
        name: range
        type: string
      - description: Return points after this telemetry ID for WS reconnection backfill
        in: query
        name: afterId
        type: integer
      - description: 'Bucketing interval: 1h (default), 6h, or 24h'
        in: query
        name: interval
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ObserverTelemetry'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Get observer telemetry history
      tags:
      - Observers
...
schemes:
- http
- https
swagger: "2.0"
...